random cleanups/cast removal/size_t love
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/gsoc_server_listing@18455 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/ServerList.h
@@ -52,7 +52,7 @@ public:
   void collectData(char *ptr, int len);
   void finalization(char *data, unsigned int length, bool good);
 
-  ServerItem* getServerAt(int index);
+  ServerItem* getServerAt(size_t index);
 
   void addServerCallback(ServerListCallback cb, void* data);
   void removeServerCallback(ServerListCallback cb, void* data);
